Some refactoring in the conversion
